% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMatrix.R
\name{cluster}
\alias{cluster}
\title{hierarchical clustering with optimal ordering}
\usage{
cluster(counts)
}
\arguments{
\item{counts}{Matrix of counts to cluster}
}
\value{
matrix with columns reordered based on clustering
}
\description{
\code{cluster} hierarchically clusters the supplied matrix by column and
orders the leaves optimally
}
\examples{
set.seed(1452)
cluster( matrix(rnorm(100), ncol = 5) )

}
