% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMatrix.R
\name{clusterMatrix}
\alias{clusterMatrix}
\title{hierarchical cluster matrix by rows or columns (or both)}
\usage{
clusterMatrix(counts, byRow = TRUE, byCol = FALSE)
}
\arguments{
\item{counts}{Matrix of counts to cluster}

\item{byRow}{logical - whether to cluster the rows of the matrix}

\item{byCol}{logical - whether to cluster the columns of the matrix}
}
\value{
reordered matrix based on clustering
}
\description{
\code{clusterMatrix} hierarchically clusters the supplied matrix by row or column
or both
}
\examples{
clusterMatrix( matrix )

clusterMatrix( matrix, byRow = FALSE, byCol = TRUE )

}
