% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_heatmap.R
\name{expr_heatmap}
\alias{expr_heatmap}
\title{expr_heatmap}
\usage{
expr_heatmap(
  expr_matrix,
  scale = NULL,
  color_scale = "viridis",
  legend_position = "left",
  row_labels = NULL
)
}
\arguments{
\item{expr_matrix}{A numeric matrix of gene expression values}

\item{scale}{character one of 'scale', 'log10'}

\item{legend_position}{character (one of 'left'[default], 'right', 'bottom', 'top')}

\item{row_labels}{character an optional character vector to allow for non-unique row labels}
}
\value{
a ggplot heatmap object
}
\description{
\code{expr_heatmap} returns a heatmap for the supplied matrix
}
\examples{
expr_heatmap( matrix, legend_position = 'right')

}
