% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotExprHeatmap.R
\name{ggplotExprHeatmap}
\alias{ggplotExprHeatmap}
\title{ggplotExprHeatmap}
\usage{
ggplotExprHeatmap(exprMatrix, legendPosition = "left")
}
\arguments{
\item{exprMatrix}{A numeric matrix of gene expression values}

\item{legendPosition}{character (one of 'left'[default], 'right', 'bottom', 'top')}
}
\value{
a ggplot heatmap object
}
\description{
\code{ggplotExprHeatmap} returns a heatmap for the supplied matrix
}
\examples{
ggplotExprHeatmap( matrix, legendPosition = 'right')

}
