% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_heatmap.R
\name{expr_heatmap}
\alias{expr_heatmap}
\title{expr_heatmap}
\usage{
expr_heatmap(
  expr_matrix,
  scale = NULL,
  colour_scale = "viridis",
  legend_position = "left",
  row_labels = NULL
)
}
\arguments{
\item{expr_matrix}{A numeric matrix of gene expression values}

\item{scale}{character one of 'scale', 'log10'}

\item{colour_scale}{character one of 'viridis', 'plasma', 'magma', 'inferno', 'byr'}

\item{legend_position}{character (one of 'left'[default], 'right', 'bottom', 'top')}

\item{row_labels}{character an optional character vector to allow for non-unique row labels}
}
\value{
a ggplot heatmap object
}
\description{
\code{expr_heatmap} returns a heatmap for the supplied matrix
}
\examples{
set.seed(1452)
means <- sample(10:100, size = 50)
log2fcs <- rnorm(50)
counts <- matrix( t(sapply(seq_len(50),
function(x){ c( rnorm(6, mean = means[x]),
rnorm(6, mean = means[x] * 2^log2fcs[x]) ) } )), nrow = 50 )
expr_heatmap( counts )

expr_heatmap( counts, legend_position = 'right')

}
